Fulfill Part 3 Reqs
Follow a TDD approach using Espresso. You can decidewhen to perform validation (i.e. dynamically as the password is entered, or only after the password is submitted). You are just required to indicate on the form whether the password is strong enough. Additional details (e.g. a red/green strength indicator bar, or providing a generic message about what is required in a password) are not required.